requirements.txt
asgi-redis==0.8.3
asgiref==0.9
autobahn==0.12.1
channels==0.9.3
daphne==0.9.1
dj-database-url==0.4.0
Django==1.9.3
djangorestframework==3.7.3
haikunator==1.0.1
msgpack-python==0.4.7
psycopg2==2.6.1
redis==2.10.5
six==1.10.0
Twisted==15.5.0
txaio==2.2.1
wheel==0.24.0
zope.interface==4.1.3
